simulator/requirements.txt
pymodbus
pyyaml